Reverted to commit 6241a1587c11cb0e26db854ed700722e8a12a581

+++ src/components/TaskEditModal.tsx
@@ -75,7 +75,7 @@ export const TaskEditModal: React.FC<TaskEditModalProps> = ({
   };
 
   const handleDelete = () => {
-    if (task) {
+    if (task && confirm('Are you sure you want to delete this task?')) {
       onDelete(task.id);
     }
   };
